Move all methods to the EventRepository
